Feat(#152) : 세션 날아감 현상 대처 및 빌드 오류 해결
