+++ src/app/page.tsx
@@ -3,15 +3,18 @@ import GenreMusicList from '@/components/main/GenreMusicList'
 import MainBanner from '@/components/main/MainBanner'
 import TopLikedBoard from '@/components/main/TopLikedBoard'
 import MusicPlayer from '@/components/player/MusicPlayer'
+import UserReSessionProvider from './UserReSessionProvider'
 
 const Home = () => {
   return (
     <>
-      <GenreMusicList />
-      <MainBanner />
-      <TopLikedBoard />
-      <MusicPlayer />
-      <Footer />
+      <UserReSessionProvider>
+        <GenreMusicList />
+        <MainBanner />
+        <TopLikedBoard />
+        <MusicPlayer />
+        <Footer />
+      </UserReSessionProvider>
     </>
   )
 }
